Divide logic in container to Form
